Change IOQ thread wait time from 5 seconds to 0.5 second
To avoid mutex contention, we wait for further requests on the same
connection rather than going to the shared thread pool to serve other
connections. The current wait time of 5 seconds is too long, causing
connections to fail if there are 2K+ connections. Waiting for half a
second should avoid the mutex contention, and it is shown to work ok
with 4K connections that we support.

Signed-off-by: Malahal Naineni <malahal@us.ibm.com>